+++ hashset.h
@@ -12,13 +12,13 @@ struct hashset {
 	size_t elt_size;
 	uint8_t *status;
 	ssize_t count;	
-	uint32_t (*hash) (const void *);
+	size_t (*hash) (const void *);
 	bool (*equals) (const void *, const void *);
 	ssize_t enlarge_threshold;	// (table size) * enlarge_factor
 };
 
 struct hashset_pos {
-	uint32_t hash;
+	size_t hash;
 	ssize_t insert;
 	ssize_t existing;
 	bool has_existing;
@@ -35,7 +35,7 @@ struct hashset_iter {
 	for ((it) = hashset_iter_make(set); hashset_iter_advance(&(it));)
 
 /* create, destroy */
-void hashset_init(struct hashset *s, uint32_t (*hash)(const void *),
+void hashset_init(struct hashset *s, size_t (*hash)(const void *),
 		  bool (*equals)(const void *, const void *),
 		  size_t elt_size);
 void hashset_init_copy(struct hashset *s, const struct hashset *src);
@@ -47,7 +47,7 @@ static inline ssize_t hashset_count(const struct hashset *s);
 static inline size_t hashset_elt_size(const struct hashset *s);
 static inline bool hashset_equals(const struct hashset *s, const void *key1,
 				  const void *key2);
-static inline uint32_t hashset_hash(const struct hashset *s, const void *key);
+static inline size_t hashset_hash(const struct hashset *s, const void *key);
 
 void *hashset_item(const struct hashset *s, const void *key);
 void *hashset_set_item(struct hashset *s, const void *key);
@@ -88,7 +88,7 @@ static inline bool hashset_equals(const struct hashset *s, const void *key1,
 	return s->equals(key1, key2);
 }
 
-static inline uint32_t hashset_hash(const struct hashset *s, const void *key)
+static inline size_t hashset_hash(const struct hashset *s, const void *key)
 {
 	return s->hash(key);
 }
